AI Battles - NetBeans project
An attempt to turn Othello AIs against each other. Each AI had different configurations and strategies. The victorious one, would theoretically be used to improve the Othello app's AI.
One of these improvements is about the "rating table", which determines how important each of the 64 Othello board tiles is, by assigning a rating to it. The greater the rating, the more important it is for a player/AI to place one of their tiles there.

Results were inconclusive at first, but I made several attempts to make something of it. I remember leaving it up and running for hours and even days! 
I don't remember much, but hey, maybe one day I will return to it.